Layout and CSS updates
Added a footer to the blog.

CSS for footers has been centralized in the common stylesheet.

CSS for headers has been relocated from the headers stylesheet to the
common stylesheet. A small styling patch was centralized out of the cgit
stylesheet. Links to the headers stylesheet are removed.

An inconsistency between the blog and git web has been fixed. The blog
has a small padding on the main element. Since cgit does not generate a
main element, this styling had to be applied to div#cgit.

An inconsistency in table layouts in git web has been fixed. Table cells
have a transparent border acting as a controlled table separation. Table
headers were lacking the same border and therefore appeared to peak out
over the column.