Linking to socials
A webring is added to the bottom of all blog posts and the home page.
This is powered by openring and stylized as a vertical list against a
light blue background (same blue as sent messages on Chat page).

My recent posts on bluesky are added to the home page. These are
stylized as received chat messages per the Chat page.

My most listened to songs (in alst 30 days) are listed on the home page.
This is an unstylized ordered list.

Minor change to the way width is applied to horizontal breaks.